+++ src/components/QuestPlaymode/Api.js
@@ -1,5 +1,5 @@
 import { BASE_URL } from '../../settings'
-import { getToken } from '../../redux/Actions/Api'
+import { getToken } from '../../api/AuthApi'
 
 export function sendTaskAttempt (questId, taskId, teamId, attemptText, callback, errorCallback) {
   fetch(`${BASE_URL}/tasks/${taskId}/attempts`, {
