+++ src/components/QuestPlaymode/Templates/Tasks/AnswerInput.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {useState} from 'react'
 import PropTypes from 'prop-types'
 import { Input } from 'antd'
 import { SendOutlined } from '@ant-design/icons'
@@ -6,11 +6,23 @@ import { SendOutlined } from '@ant-design/icons'
 export default function AnswerInput (props) {
   const ANSWER_ACCEPTED_STATUS = 'accepted'
 
+  const [t, setT] = useState(new Date())
+
   const getPlaceholder = () => {
     if (props.manualVerificationEnabled) {
       return 'Проверка модератором'
     }
     return 'Автоматическая проверка'
+  }
+
+  function isOkTimedelta(td) {
+    const delta = (td - t)
+    if (delta >= 5000) {
+      setT(new Date())
+      return true
+    } else {
+      alert("Подожди еще немного"); return false
+    }
   }
 
   if (props.answerStatus === ANSWER_ACCEPTED_STATUS) {
@@ -30,7 +42,9 @@ export default function AnswerInput (props) {
         disabled={false}
         placeholder={getPlaceholder()}
         enterButton=<SendOutlined />
-        onSearch={(value) => { if (value !== '') {props.sendAnswer(value)} }}
+        onSearch={(value) => { if (value !== '' && isOkTimedelta(new Date())) {
+          props.sendAnswer(value)}
+        }}
       />
     </div>)
 }
