+++ src/components/shared/Header/Header.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useHistory } from 'react-router-dom'
 import { Divider, PageHeader } from 'antd'
 import { connect } from 'react-redux'
-import { logout } from '../../../redux/Actions/Api'
+import { logout } from '../../../api/AuthApi'
 import UserButton from './UserButton/UserButton'
 import QuestspaceIcon from '../Icons/QuestspaceIcon'
 import './Header.css'
